backend/requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and AI dependencies
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1
safetensors==0.4.0
scipy==1.11.4

# Image processing dependencies
Pillow==10.1.0
opencv-python==4.8.1.78

# Utility dependencies
numpy==1.24.3
requests==2.31.0
huggingface-hub==0.19.4
tokenizers==0.15.0

# Optional GPU acceleration (uncomment if using CUDA)
# torch==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.16.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Development dependencies (optional)
python-dotenv==1.0.0 